+++ tests/testthat/test-seq_linter.R
@@ -232,3 +232,90 @@ test_that("seq_len(length(x)) should be seq_along(x)", {
   expect_lint("seq_len(length(x))", "seq_along", linter)
   expect_lint("seq_len(length(foo(x)))", "seq_along", linter)
 })
+
+test_that("finds seq(1, n) and seq(from = 1, to = n) expressions", {
+  linter <- seq_linter()
+  lint_msg <- function(want, got) rex::rex("Use ", want, " instead of ", got)
+
+  expect_lint(
+    "seq(1, 10)",
+    lint_msg("seq_len(10)", "seq(1, 10)"),
+    linter
+  )
+
+  expect_lint(
+    "base::seq(1, 10)",
+    lint_msg("seq_len(10)", "seq(1, 10)"),
+    linter
+  )
+
+  expect_lint(
+    "seq(1L, 10L)",
+    lint_msg("seq_len(10L)", "seq(1L, 10L)"),
+    linter
+  )
+
+  expect_lint(
+    "seq(1, n)",
+    lint_msg("seq_len(n)", "seq(1, n)"),
+    linter
+  )
+
+  expect_lint(
+    "seq(1, length(x))",
+    lint_msg("seq_along(...)", "seq(1, length(...))"),
+    linter
+  )
+
+  expect_lint(
+    "seq(1, nrow(x))",
+    lint_msg("seq_len(nrow(...))", "seq(1, nrow(...))"),
+    linter
+  )
+
+  expect_lint(
+    "seq(from = 1, to = 10)",
+    lint_msg("seq_len(10)", "seq(1, 10)"),
+    linter
+  )
+
+  expect_lint(
+    "seq(from = 1L, to = length(x))",
+    lint_msg("seq_along(...)", "seq(1L, length(...))"),
+    linter
+  )
+
+  expect_lint(
+    "seq(to = 10, from = 1)",
+    lint_msg("seq_len(10)", "seq(1, 10)"),
+    linter
+  )
+
+  expect_lint(
+    "seq(to = length(x), from = 1L)",
+    lint_msg("seq_along(...)", "seq(1L, length(...))"),
+    linter
+  )
+
+  expect_lint(
+    "seq(1, to = 10)",
+    lint_msg("seq_len(10)", "seq(1, 10)"),
+    linter
+  )
+
+  expect_lint(
+    "seq(to = 10, 1)",
+    lint_msg("seq_len(10)", "seq(1, 10)"),
+    linter
+  )
+
+  # Complex expressions or other arguments
+  expect_no_lint("seq(2, 10)", linter)
+  expect_no_lint("seq(1, 10, by = 2)", linter)
+  expect_no_lint("seq(1, 10, length.out = 5)", linter)
+  expect_no_lint("seq(1, 10, 2)", linter)
+  expect_no_lint("seq(from = 2, to = 10)", linter)
+  expect_no_lint("seq(from = 1, to = 10, by = 2)", linter)
+})
+
+